Inline int.floor_div and int.modulo
With the information we have (that the divisor is never zero) we can
do a slightly more optimized approach than the standard library can.

I'm not sure what the difference is and I should really benchmark them
but it was flagged by it being quite expensive in a real work project
which was doing a lot of calculations with dates.

I've also changed from doing 'a * b < 0' to see if either a or b is
less than zero (but not the other) to doing the comparisons explicitly
as I perhaps incorrectly feel like it would be few CPU cycles.